+++ Sources/Sentry/PrivateSentrySDKOnly.mm
@@ -321,6 +321,12 @@ static BOOL _framesTrackingMeasurementHybridSDKMode = NO;
 }
 
 #if SENTRY_TARGET_REPLAY_SUPPORTED
+
++ (UIView *)sessionReplayMaskingOverlay:(id<SentryRedactOptions>)options
+{
+    return [[SentryMaskingPreviewView alloc] initWithRedactOptions:options];
+}
+
 + (nullable SentrySessionReplayIntegration *)getReplayIntegration
 {