+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ### Features
 
 - Add `showMaskPreview` to `SentrySDK.replay` api to debug replay masking (#4761)
+- Session replay masking preview for SwiftUI (#4737)
 
 ## 8.44.0-beta.1